feat: complete open search dialog and display items on search
Added the feature that allows the user to open the search dialog by
pressing a + button for the associated meal. When the user searches an
item it pops up within the searchlist of the dialog.

Also added a route for /dashboard and changed auth.js to redirect to the correct page upon signup/login.
Also initialize users' calorie_goal to 0 upon signing up.

Progress made towards GH Issue #5