feat: add calorie bar graph svg (#16)
Added the html needed for the actual bar graphs.

NOTE
There is a minor issue with the current iteration of the bar graphs in
that due to the use of stroke-endcap="round" a bar that is "1%" full
will always display a circle instead of being 1% of the allocated track
area. This is a quirk of using endcaps, look into making the track a
<path> that masks a non-nonendcapped line.